+++ src/lib/isFromKeyboard.ts
@@ -1,0 +1,12 @@
+import whatInput from 'what-input'
+
+export default class IsFromKeyboard {
+  static state = () => {
+    const isFromKeyboard = whatInput.ask() === 'keyboard'
+    return { isFromKeyboard }
+  }
+
+  static initial = { isFromKeyboard: false }
+
+  static propertyName = 'isFromKeyboard'
+}